refactor(ai-dock): the metric panel uses the shared pending-query hook
PromGraphCpt carried its own copy of usePendingQuery: the same
AbortController, the same settled guard, the same pendingRef bookkeeping.
The copy had already started to drift — services.ts re-declared
QueryRequest, so Table and Graph imported a second definition of a type
the MySQL and Doris panels take from the hook.

`begin` stays adjacent to `submit`, the way the form-backed factory keeps
it next to `commit`: the request and the refresh have to reach the fetcher
in one render.

The dock's `className` prop went with it. No caller ever passed one.
